refactor: move events to events/events.py
