+++ anilibria/api/gateway/events/events.py
@@ -1,0 +1,16 @@
+
+
+@define()
+class TitleSerie:
+    """
+    Модель для ивента `on_title_serie` и подписок.
+
+    .. code-block:: python
+
+      @client.event
+      async def on_title_serie(event: TitleSerie):
+          ...
+    """
+
+    title: Title = field(converter=convert(Title))
+    episode: Serie = field()  # Нет конвертера, так как передаётся из другой модели
